Fix POP + "Toggle Slow Animations in Frontmost App"
Turns out, it's actually just a float. 🎉

+++ pop/POPAnimationExtras.mm
@@ -17,7 +17,7 @@
 #endif
 
 #if TARGET_IPHONE_SIMULATOR
-UIKIT_EXTERN CGFloat UIAnimationDragCoefficient(); // UIKit private drag coeffient, use judiciously
+UIKIT_EXTERN float UIAnimationDragCoefficient(); // UIKit private drag coeffient, use judiciously
 #endif
 
 CGFloat POPAnimationDragCoefficient()
